Switch to C on the backend.

+++ src/backend/backend.ts
@@ -4,5 +4,5 @@ export interface BackendContext {
   indentLevel: int;
   append: (value: string) => void;
   prepend: (value: string) => void;
-  remove: (count: u64) => void;
+  remove: (count: int) => void;
 }
